refactor relatedChannels
